fix(SkyBlockTime): definition of a different class method breaking tests of object classes
