+++ src/main/java/at/hannibal2/skyhanni/features/misc/InGameDateDisplay.kt
@@ -4,7 +4,7 @@ import at.hannibal2.skyhanni.SkyHanniMod
 import at.hannibal2.skyhanni.events.GuiRenderEvent
 import at.hannibal2.skyhanni.events.LorenzTickEvent
 import at.hannibal2.skyhanni.utils.LorenzUtils
-import at.hannibal2.skyhanni.utils.LorenzUtils.formatted
+import at.hannibal2.skyhanni.utils.NEUSkyblockTime.formatted
 import at.hannibal2.skyhanni.utils.RenderUtils.renderString
 import io.github.moulberry.notenoughupdates.util.SkyBlockTime
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent
